New Node wws-token-viewer
A dedicated token to check the status of the outh token.
Based on a refresh period configured [0 - 60] the token is getting
refreshed automatically.

This is useful for call which are not executed within the grace period
of an expired user token (max 2h).

By this it will be assured that the token will be refreshed
automatically.